Warning Do not Pull this one First
---------------------------WARNING-------------------------------------
Do not pull this one even if you see it, because it will delete a bunch of old scenes, scripts, prefeb, material and other asserts. 
That's because I want to make everything in the file easy to be located, I've made a backup on the Cloud. 
And if you want to pull,
MAKE  SURE YOU MADE A BACK UP ALREADY.
for the old file.
-----------------------UPDATE INFORM-----------------------------------------
This one mainly combine Daniel's script to those 3 different puzzle scenes and put all 3 into one scene.
Right now there is only one scene, each time player pass a certain symbol, Camera will move to offer a batter angle for player for the rest of the game. Also, in some situation, the Directional light also need to be changed. However, due to the reason that rotation is 100000 times harder than position, directional light will immediately rotate to that angle instead of smoothly rotate to there.
Also the old "load scene" shouldn't function now. To do that, we might create different classes inhereted from one for different kind of situations