Loading Scene Bug Fixed
The loading scenes bug(takes long time to load) is fixed now, The loading scene should takes fewer second.
If you're interested, the problem is:
Because I use Event System to let the level manager to load the scene. I need to register the loading scene function into the a Directionary inside EventManager. However, EventManager is a singleton which won't be destroy when loading another scene, but level manager will try register the function again. And because the level manager in different scene is different, thus it will successfully register the loading scene function. And the result is, it will load the scene multiple times and make it super slow!!!